requirements.txt
bert-extractive-summarizer
spacy
transformers
torch
tensorflow
streamlit